+++ core/src/commonMain/kotlin/com/deezer/caupain/model/Repository.kt
@@ -34,6 +34,7 @@ import io.ktor.http.HttpHeaders
 import io.ktor.http.URLBuilder
 import kotlin.io.encoding.Base64
 import kotlin.io.encoding.ExperimentalEncodingApi
+import kotlin.jvm.JvmOverloads
 
 /**
  * Maven repository
@@ -45,9 +46,36 @@ import kotlin.io.encoding.ExperimentalEncodingApi
 public class Repository(
     public val url: String,
     public val user: String?,
-    public val password: String?
+    public val password: String?,
+    public val componentFilter: ComponentFilter? = null
 ) : Serializable {
-    public constructor(url: String) : this(url, null, null)
+    @JvmOverloads
+    public constructor(
+        url: String,
+        componentFilter: ComponentFilter? = null
+    ) : this(url, null, null, componentFilter)
+
+    /**
+     * Checks if the given [Dependency] is accepted by this repository.
+     */
+    public operator fun contains(dependency: Dependency): Boolean {
+        return componentFilter == null || componentFilter.accepts(dependency)
+    }
+
+    /**
+     * Returns a repository with the given [ComponentFilter] applied.
+     *
+     * @see ComponentFilterBuilder.includes
+     * @see ComponentFilterBuilder.excludes
+     */
+    public inline fun withComponentFilter(builder: ComponentFilterBuilder.() -> Unit): Repository {
+        return Repository(
+            url = url,
+            user = user,
+            password = password,
+            componentFilter = buildComponentFilter(builder)
+        )
+    }
 
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
@@ -58,6 +86,7 @@ public class Repository(
         if (url != other.url) return false
         if (user != other.user) return false
         if (password != other.password) return false
+        if (componentFilter != other.componentFilter) return false
 
         return true
     }
@@ -66,6 +95,7 @@ public class Repository(
         var result = url.hashCode()
         result = 31 * result + (user?.hashCode() ?: 0)
         result = 31 * result + (password?.hashCode() ?: 0)
+        result = 31 * result + (componentFilter?.hashCode() ?: 0)
         return result
     }
 
@@ -75,6 +105,96 @@ public class Repository(
 
     public companion object {
         private const val serialVersionUID = 1L
+    }
+}
+
+/**
+ * Represents a filter for components in a repository. If used in the [Repository], only the [Dependency]
+ * that match the filter will be checked against the given [Repository].
+ */
+public interface ComponentFilter : Serializable {
+    /**
+     * Checks if the given [Dependency] is accepted by this filter.
+     *
+     * @param dependency The dependency to check.
+     * @return `true` if the dependency is accepted, `false` otherwise.
+     */
+    public fun accepts(dependency: Dependency): Boolean
+}
+
+/**
+ * Builder for [ComponentFilter]. If only excludes are defined, then all dependencies except those
+ * excluded will be accepted. If only includes are defined, then only those dependencies will be accepted.
+ * If both are defined, then only those dependencies that are included and not excluded will be accepted.
+ */
+public class ComponentFilterBuilder {
+    private val includes = mutableListOf<PackageSpec>()
+    private val excludes = mutableListOf<PackageSpec>()
+
+    /**
+     * Excludes a dependency from this repository. If name is null, group is used as a glob, with the following rules:
+     * - `*`: wildcard that matches zero, one or multiple characters, other than `.`
+     * - `**`: Wildcard that matches zero, one or multiple packages. For example, `**.sub.name` matches
+     * `com.example.sub.name`, `com.example.sub.sub.name`. `**` must be either preceded by `.` or be at
+     * the beginning of the glob. `**` must be either followed by `.` or be at the end of the glob.
+     * If the glob only consist of a `**`, it will be a match for everything.
+     *
+     * @property group The group to exclude. If `name` is null, then this is interpreted as a glob
+     * @property name The name to exclude. If null, all libraries in the group are excluded.
+     */
+    @JvmOverloads
+    public fun exclude(group: String, name: String? = null): ComponentFilterBuilder {
+        excludes.add(PackageSpec(group, name))
+        return this
+    }
+
+    /**
+     * Includes a dependency in this repository. If name is null, group is used as a glob, with the following rules:
+     * - `*`: wildcard that matches zero, one or multiple characters, other than `.`
+     * - `**`: Wildcard that matches zero, one or multiple packages. For example, `**.sub.name` matches
+     * `com.example.sub.name`, `com.example.sub.sub.name`. `**` must be either preceded by `.` or be at
+     * the beginning of the glob. `**` must be either followed by `.` or be at the end of the glob.
+     * If the glob only consist of a `**`, it will be a match for everything.
+     *
+     * @property group The group to include. If `name` is null, then this is interpreted as a glob
+     * @property name The name to included. If null, all libraries in the group are included.
+     */
+    @JvmOverloads
+    public fun include(group: String, name: String? = null): ComponentFilterBuilder {
+        includes.add(PackageSpec(group, name))
+        return this
+    }
+
+    /**
+     * Builds the [ComponentFilter] with the current includes and excludes.
+     */
+    public fun build(): ComponentFilter {
+        return DefaultComponentFilter(includes.toList(), excludes.toList())
+    }
+}
+
+/**
+ * Builds a [ComponentFilter] using the given builder function.
+ *
+ * @see ComponentFilterBuilder.includes
+ * @see ComponentFilterBuilder.excludes
+ */
+public inline fun buildComponentFilter(builder: ComponentFilterBuilder.() -> Unit): ComponentFilter {
+    return ComponentFilterBuilder().apply(builder).build()
+}
+
+internal data class DefaultComponentFilter(
+    private val includes: List<PackageSpec>,
+    private val excludes: List<PackageSpec>,
+) : ComponentFilter {
+    override fun accepts(dependency: Dependency): Boolean {
+        if (excludes.isNotEmpty()) {
+            if (excludes.any { it.matches(dependency) }) return false
+        }
+        if (includes.isNotEmpty()) {
+            if (includes.none { it.matches(dependency) }) return false
+        }
+        return true
     }
 }