+++ gradle-plugin/src/main/java/com/deezer/caupain/plugin/RepositoryHandler.kt
@@ -26,17 +26,26 @@
 
 package com.deezer.caupain.plugin
 
+import com.deezer.caupain.model.ComponentFilter
+import com.deezer.caupain.model.ComponentFilterBuilder
+import com.deezer.caupain.model.Dependency
+import com.deezer.caupain.model.Dependency.Library
+import com.deezer.caupain.model.Dependency.Plugin
 import com.deezer.caupain.model.Repository
+import com.deezer.caupain.model.buildComponentFilter
 import org.gradle.api.Action
+import org.gradle.api.artifacts.ModuleIdentifier
+import org.gradle.api.artifacts.component.ModuleComponentIdentifier
 import org.gradle.api.artifacts.dsl.RepositoryHandler
 import org.gradle.api.artifacts.repositories.MavenArtifactRepository
 import org.gradle.api.artifacts.repositories.PasswordCredentials
 import org.gradle.api.internal.GradleInternal
+import org.gradle.api.internal.artifacts.repositories.ArtifactResolutionDetails
+import org.gradle.api.internal.artifacts.repositories.ContentFilteringRepository
 import org.gradle.api.invocation.Gradle
 import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.ListProperty
 import org.gradle.api.provider.Provider
-import org.gradle.api.tasks.Input
 import org.gradle.internal.artifacts.repositories.AuthenticationSupportedInternal
 import org.gradle.kotlin.dsl.listProperty
 import java.util.Optional
@@ -55,7 +64,6 @@ abstract class RepositoryHandler @Inject constructor(
     /**
      * Libraries repositories to check for updates.
      */
-    @get:Input
     val libraries: ListProperty<Repository> = objects.listProperty<Repository>().convention(
         gradle.dependenciesRepositoryHandler.toRepositories(objects)
     )
@@ -63,7 +71,6 @@ abstract class RepositoryHandler @Inject constructor(
     /**
      * Plugin repositories
      */
-    @get:Input
     val plugins: ListProperty<Repository> = objects.listProperty<Repository>().convention(
         gradle.pluginsRepositoryHandler.toRepositories(objects)
     )
@@ -92,8 +99,14 @@ private fun RepositoryHandler.toRepositories(objects: ObjectFactory): Provider<L
             val url = repository.url.toString()
             val credentials =
                 (repository as? AuthenticationSupportedInternal)?.configuredCredentials
+            val filter = (repository as? ContentFilteringRepository)?.contentFilter
             if (credentials == null) {
-                repositoriesProvider.add(Repository(url = url))
+                repositoriesProvider.add(
+                    Repository(
+                        url = url,
+                        componentFilter = filter?.let(::ComponentFilterAdapter)
+                    )
+                )
             } else {
                 repositoriesProvider.add(
                     credentials
@@ -105,7 +118,8 @@ private fun RepositoryHandler.toRepositories(objects: ObjectFactory): Provider<L
                             Repository(
                                 url = url,
                                 user = passwordCredentials?.username,
-                                password = passwordCredentials?.password
+                                password = passwordCredentials?.password,
+                                componentFilter = filter?.let(::ComponentFilterAdapter)
                             )
                         }
                 )
@@ -115,10 +129,86 @@ private fun RepositoryHandler.toRepositories(objects: ObjectFactory): Provider<L
     return repositoriesProvider
 }
 
+private class ComponentFilterAdapter(private val filter: Action<in ArtifactResolutionDetails>) :
+    ComponentFilter {
+
+    override fun accepts(dependency: Dependency): Boolean {
+        if (dependency.group == null || dependency.name == null) return false
+        return ArtifactResolutionDetailsAdapter(dependency)
+            .apply { filter.execute(this) }
+            .isFound
+    }
+
+    private class ModuleIdentifierAdapter(private val dependency: Dependency) : ModuleIdentifier {
+        override fun getGroup(): String = requireNotNull(dependency.group)
+
+        override fun getName(): String = requireNotNull(dependency.name)
+    }
+
+    private class ModuleComponentIdentifierAdapter(private val dependency: Dependency) :
+        ModuleComponentIdentifier {
+        private val identifier = ModuleIdentifierAdapter(dependency)
+
+        override fun getDisplayName(): String = dependency.moduleId
+
+        override fun getGroup(): String = identifier.group
+
+        override fun getModule(): String = identifier.name
+
+        override fun getVersion(): String = requireNotNull(dependency.version).toString()
+
+        override fun getModuleIdentifier(): ModuleIdentifier = identifier
+    }
+
+    private class ArtifactResolutionDetailsAdapter(
+        private val dependency: Dependency
+    ) : ArtifactResolutionDetails {
+
+        var isFound = true
+            private set
+
+        override fun getModuleId(): ModuleIdentifier = ModuleIdentifierAdapter(dependency)
+
+        override fun getComponentId(): ModuleComponentIdentifier? =
+            if (dependency.version == null) {
+                null
+            } else {
+                ModuleComponentIdentifierAdapter(dependency)
+            }
+
+        override fun isVersionListing(): Boolean = dependency.version != null
+
+        override fun notFound() {
+            isFound = false
+        }
+    }
+
+    companion object {
+        private val Dependency.group: String?
+            get() = when (this) {
+                is Library -> group
+                is Plugin -> id
+            }
+
+        private val Dependency.name: String?
+            get() = when (this) {
+                is Library -> name
+                is Plugin -> "$id.gradle.plugin"
+            }
+    }
+}
+
 class RepositoryCategoryHandler internal constructor(private val listProperty: ListProperty<Repository>) {
 
     fun repository(repository: Repository) {
         listProperty.add(repository)
+    }
+
+    fun repository(
+        repository: Repository,
+        configureComponentFilter: Action<ComponentFilterBuilder>
+    ) {
+        listProperty.add(repository.withComponentFilter { configureComponentFilter.execute(this) })
     }
 
     /**
@@ -128,10 +218,39 @@ class RepositoryCategoryHandler internal constructor(private val listProperty: L
         listProperty.add(Repository(url))
     }
 
+    fun repository(url: String, configureComponentFilter: Action<ComponentFilterBuilder>) {
+        listProperty.add(
+            Repository(
+                url = url,
+                componentFilter = buildComponentFilter {
+                    configureComponentFilter.execute(this)
+                }
+            )
+        )
+    }
+
     /**
      * Adds a repository with authentication
      */
     fun repository(url: String, user: String, password: String) {
         listProperty.add(Repository(url, user, password))
+    }
+
+    fun repository(
+        url: String,
+        user: String,
+        password: String,
+        configureComponentFilter: Action<ComponentFilterBuilder>
+    ) {
+        listProperty.add(
+            Repository(
+                url = url,
+                user = user,
+                password = password,
+                componentFilter = buildComponentFilter {
+                    configureComponentFilter.execute(this)
+                }
+            )
+        )
     }
 }